+++ src/Symfony/Component/Messenger/Bridge/Beanstalkd/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `bury_on_reject` option to bury failed messages instead of deleting them
+ * Add `BeanstalkdPriorityStamp` option to allow setting the message priority
 
 5.2.0
 -----